<Error> : <Error 처리 함수 리팩토링>
ref : Application 에 있던 에러를 처리하는 함수들을 errorManage 클래스로 분리 하였고 실행을 하는 클래스인 RunGame 을 따로 분리하였음.
